requirements.txt
asgiref==3.9.1
certifi==2025.8.3
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
elastic-transport==9.1.0
elasticsearch==9.1.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
